Revert "Remove calculation of inverse sqrt of covariance matrix"
This reverts commit 968c6b00e98b8f639b31a7dbd32f4cc8e15a4a63.

Further testing has revealed a significant performance regression for very
large lambda values, as are encountered commonly when using IPOP and BIPOP.
This regression can be seen even for large N (up to N=100), which suggests
that the cost of an additional clone per sampled point outweighs the cost of
two matrix multiplications per eigen update, except for when N is very large
(N > 100). However, the performance gain from the reverted commit for large N
seems to be minor in any case, so it should be safe to remove unless such
large N values turn out to be a common use case.